모든 페이지 다듬기, 식단 리스트 Card로 대체

+++ project1/lib/proflie_set_page.dart
@@ -7,7 +7,8 @@ import 'package:project1/login_page.dart';
 // 프로필 설정 페이지
 
 class ProflieSetPage extends StatefulWidget {
-  ProflieSetPage({super.key});
+  final String age, height;
+  ProflieSetPage(this.age, this.height, {super.key});
 
   @override
   State<ProflieSetPage> createState() => _ProflieSetPageState();
@@ -16,8 +17,12 @@ class ProflieSetPage extends StatefulWidget {
 class _ProflieSetPageState extends State<ProflieSetPage> {
   final storage = FlutterSecureStorage();
   final _form = GlobalKey<FormState>();
-  late String _age, _height, _markWeight;
-  bool _markEnable = false;
+  late String _age, _height;
+
+  @override
+  void initState() {
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -25,43 +30,29 @@ class _ProflieSetPageState extends State<ProflieSetPage> {
         appBar: AppBar(
           title: const Text("설정"),
         ),
-        body: SingleChildScrollView(
-            child: Padding(
-          padding: const EdgeInsets.all(30),
-          child: Form(
-            key: _form,
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Row(
-                  children: [
-                    Flexible(child: ageInput()),
-                    Container(width: 10),
-                    Flexible(child: heightInput()),
-                  ],
-                ),
-                Container(height: 10),
-                Row(
-                  children: [
-                    Flexible(child: markWeightInput()),
-                    Container(width: 10),
-                    Switch(
-                      value: _markEnable,
-                      onChanged: (value) {
-                        setState(() {
-                          _markEnable = value;
-                        });
-                      },
-                    )
-                  ],
-                ),
-                Container(height: 10),
-                profileSubmitButton(),
-                logoutButton(),
-              ],
+        body: Center(
+          child: SingleChildScrollView(
+              child: Padding(
+            padding: const EdgeInsets.all(30),
+            child: Form(
+              key: _form,
+              child: Column(
+                children: [
+                  Row(
+                    children: [
+                      Flexible(child: ageInput()),
+                      Container(width: 10),
+                      Flexible(child: heightInput()),
+                    ],
+                  ),
+                  Container(height: 10),
+                  profileSubmitButton(),
+                  logoutButton(),
+                ],
+              ),
             ),
-          ),
-        )));
+          )),
+        ));
   }
 
   Widget ageInput() {
@@ -83,6 +74,7 @@ class _ProflieSetPageState extends State<ProflieSetPage> {
           errorStyle: TextStyle(fontSize: 0)),
       keyboardType: TextInputType.number,
       inputFormatters: [FilteringTextInputFormatter.digitsOnly],
+      initialValue: widget.age,
     );
   }
 
@@ -105,29 +97,7 @@ class _ProflieSetPageState extends State<ProflieSetPage> {
           errorStyle: TextStyle(fontSize: 0)),
       keyboardType: TextInputType.number,
       inputFormatters: [FilteringTextInputFormatter.digitsOnly],
-    );
-  }
-
-  Widget markWeightInput() {
-    return TextFormField(
-      validator: (value) {
-        if (value!.isEmpty) {
-          return "";
-        } else {
-          return null;
-        }
-      },
-      onSaved: (newValue) {
-        _markWeight = newValue as String;
-      },
-      enabled: _markEnable,
-      decoration: InputDecoration(
-          border: OutlineInputBorder(),
-          labelText: "목표체중",
-          suffixText: "kg",
-          errorStyle: TextStyle(fontSize: 0)),
-      keyboardType: TextInputType.number,
-      inputFormatters: [FilteringTextInputFormatter.digitsOnly],
+      initialValue: widget.height,
     );
   }
 
@@ -136,24 +106,47 @@ class _ProflieSetPageState extends State<ProflieSetPage> {
         onPressed: () {
           if (_form.currentState!.validate()) {
             _form.currentState!.save();
-            Map<String, dynamic> profileMap = {
-              "age": _age,
-              "height": _height,
-              "markWeight": _markWeight
-            };
+            Map<String, dynamic> profileMap = {"age": _age, "height": _height};
             addProfileFunc(profileMap);
+            Navigator.of(context).pop();
           }
         },
-        child: Text("수정"));
+        child: Text("적용"));
   }
 
   Widget logoutButton() {
     return ElevatedButton(
-        onPressed: () async {
-          await storage.delete(key: "uid");
-          await storage.write(key: "loginState", value: "false");
-          Navigator.of(context)
-              .push(MaterialPageRoute(builder: (context) => LoginPage()));
+        onPressed: () {
+          showDialog(
+            context: context,
+            builder: (context) {
+              return AlertDialog(
+                title: Text("Logout"),
+                content: const Text("로그아웃 하시겠습니까?"),
+                actions: [
+                  Row(
+                    mainAxisAlignment: MainAxisAlignment.spaceAround,
+                    children: [
+                      FilledButton(
+                          onPressed: () async {
+                            await storage.delete(key: "uid");
+                            await storage.write(
+                                key: "loginState", value: "false");
+                            Navigator.of(context).push(MaterialPageRoute(
+                                builder: (context) => LoginPage()));
+                          },
+                          child: Text("확인")),
+                      TextButton(
+                          onPressed: () {
+                            Navigator.pop(context);
+                          },
+                          child: Text("취소"))
+                    ],
+                  )
+                ],
+              );
+            },
+          );
         },
         child: Text("로그아웃"));
   }
